Added executor + refactor
