+++ internal/runner/options.go
@@ -7,15 +7,6 @@ import (
 	"github.com/projectdiscovery/gologger"
 )
 
-// DefaultResolvers contains the list of resolvers known to be trusted.
-var DefaultResolvers = []string{
-	"1.1.1.1:53", // Cloudflare
-	"1.0.0.1:53", // Cloudflare
-	"8.8.8.8:53", // Google
-	"8.8.4.4:53", // Google
-	"9.9.9.9:53", // Quad9
-}
-
 // Options contains the configuration options for tuning
 // the template requesting process.
 type Options struct {
