+++ internal/runner/runner.go
@@ -165,59 +165,7 @@ func (r *Runner) processTemplateWithList(template *templates.Template, request i
 func (r *Runner) sendRequest(template *templates.Template, request interface{}, URL string, writer *bufio.Writer, httpclient *retryablehttp.Client, dnsclient *retryabledns.Client) {
 	switch request.(type) {
 	case *requests.HTTPRequest:
-		if !isURL(URL) {
-			break
-		}
-		httpRequest := request.(*requests.HTTPRequest)
-
-		// Compile each request for the template based on the URL
-		compiledRequest, err := httpRequest.MakeHTTPRequest(URL)
-		if err != nil {
-			gologger.Warningf("[%s] Could not make request %s: %s\n", template.ID, URL, err)
-			return
-		}
-
-		// Send the request to the target servers
-		for _, req := range compiledRequest {
-			resp, err := httpclient.Do(req)
-			if err != nil {
-				if resp != nil {
-					resp.Body.Close()
-				}
-				gologger.Warningf("[%s] Could not send request %s: %s\n", template.ID, URL, err)
-				return
-			}
-
-			data, err := ioutil.ReadAll(resp.Body)
-			if err != nil {
-				io.Copy(ioutil.Discard, resp.Body)
-				resp.Body.Close()
-				gologger.Warningf("[%s] Could not read body %s: %s\n", template.ID, URL, err)
-				continue
-			}
-			resp.Body.Close()
-
-			body := unsafeToString(data)
-
-			var headers string
-			for _, matcher := range httpRequest.Matchers {
-				// Only build the headers string if the matcher asks for it
-				part := matcher.GetPart()
-				if part == matchers.AllPart || part == matchers.HeaderPart && headers == "" {
-					headers = headersToString(resp.Header)
-				}
-
-				// Check if the matcher matched
-				if matcher.Match(resp, body, headers) {
-					// If there is an extractor, run it.
-					var extractorResults []string
-					for _, extractor := range httpRequest.Extractors {
-						part := extractor.GetPart()
-						if part == extractors.AllPart || part == extractors.HeaderPart && headers == "" {
-							headers = headersToString(resp.Header)
-						}
-						extractorResults = append(extractorResults, extractor.Extract(body, headers)...)
-					}
+	
 
 					// All the matchers matched, print the output on the screen
 					output := buildOutputHTTP(template, req, extractorResults, matcher)
@@ -232,40 +180,6 @@ func (r *Runner) sendRequest(template *templates.Template, request interface{}, 
 			}
 		}
 	case *requests.DNSRequest:
-		// eventually extracts dns from url
-		var domain string = URL
-		if isURL(URL) {
-			domain = extractDomain(URL)
-		}
-
-		dnsRequest := request.(*requests.DNSRequest)
-
-		// Compile each request for the template based on the URL
-		compiledRequest, err := dnsRequest.MakeDNSRequest(domain)
-		if err != nil {
-			gologger.Warningf("[%s] Could not make request %s: %s\n", template.ID, domain, err)
-			return
-		}
-
-		// Send the request to the target servers
-		resp, err := dnsclient.Do(compiledRequest)
-		if err != nil {
-			gologger.Warningf("[%s] Could not send request %s: %s\n", template.ID, domain, err)
-			return
-		}
-
-		for _, matcher := range dnsRequest.Matchers {
-			// Check if the matcher matched
-			if !matcher.MatchDNS(resp) {
-				return
-			}
-		}
-
-		// If there is an extractor, run it.
-		var extractorResults []string
-		for _, extractor := range dnsRequest.Extractors {
-			extractorResults = append(extractorResults, extractor.ExtractDNS(resp.String())...)
-		}
 
 		// All the matchers matched, print the output on the screen
 		output := buildOutputDNS(template, domain, extractorResults)
@@ -277,118 +191,4 @@ func (r *Runner) sendRequest(template *templates.Template, request interface{}, 
 			r.outputMutex.Unlock()
 		}
 	}
-}
-
-// buildOutputHTTP builds an output text for writing results
-func buildOutputHTTP(template *templates.Template, req *retryablehttp.Request, extractorResults []string, matcher *matchers.Matcher) string {
-	builder := &strings.Builder{}
-	builder.WriteRune('[')
-	builder.WriteString(template.ID)
-	if len(matcher.Name) > 0 {
-		builder.WriteString(":")
-		builder.WriteString(matcher.Name)
-	}
-	builder.WriteString("] ")
-
-	// Escape the URL by replacing all % with %%
-	URL := req.URL.String()
-	escapedURL := strings.Replace(URL, "%", "%%", -1)
-	builder.WriteString(escapedURL)
-
-	// If any extractors, write the results
-	if len(extractorResults) > 0 {
-		builder.WriteString(" [")
-		for i, result := range extractorResults {
-			builder.WriteString(result)
-			if i != len(extractorResults)-1 {
-				builder.WriteRune(',')
-			}
-		}
-		builder.WriteString("]")
-	}
-	builder.WriteRune('\n')
-
-	return builder.String()
-}
-
-// buildOutput builds an output text for writing results
-func buildOutputDNS(template *templates.Template, domain string, extractorResults []string) string {
-	builder := &strings.Builder{}
-	builder.WriteRune('[')
-	builder.WriteString(template.ID)
-	builder.WriteString("] [dns] ")
-
-	builder.WriteString(domain)
-
-	// If any extractors, write the results
-	if len(extractorResults) > 0 {
-		builder.WriteString(" [")
-		for i, result := range extractorResults {
-			builder.WriteString(result)
-			if i != len(extractorResults)-1 {
-				builder.WriteRune(',')
-			}
-		}
-		builder.WriteString("]")
-	}
-	builder.WriteRune('\n')
-
-	return builder.String()
-}
-
-// makeHTTPClient creates a HTTP client with configurable redirect field
-func (r *Runner) makeHTTPClientByRequest(request interface{}) *retryablehttp.Client {
-
-	redirects := false
-	maxRedirects := 0
-	// Request is HTTP
-	if httpRequest, ok := request.(requests.HTTPRequest); ok {
-		redirects = httpRequest.Redirects
-		maxRedirects = httpRequest.MaxRedirects
-	}
-
-	retryablehttpOptions := retryablehttp.DefaultOptionsSpraying
-	retryablehttpOptions.RetryWaitMax = 10 * time.Second
-	retryablehttpOptions.RetryMax = r.options.Retries
-
-	// Create the HTTP Client
-	client := retryablehttp.NewWithHTTPClient(&http.Client{
-		Transport: &http.Transport{
-			MaxIdleConnsPerHost: -1,
-			TLSClientConfig: &tls.Config{
-				Renegotiation:      tls.RenegotiateOnceAsClient,
-				InsecureSkipVerify: true,
-			},
-			DisableKeepAlives: true,
-		},
-		Timeout: time.Duration(r.options.Timeout) * time.Second,
-		CheckRedirect: func(_ *http.Request, requests []*http.Request) error {
-			if !redirects {
-				return http.ErrUseLastResponse
-			}
-			if maxRedirects == 0 {
-				if len(requests) > 10 {
-					return http.ErrUseLastResponse
-				}
-				return nil
-			}
-			if len(requests) > maxRedirects {
-				return http.ErrUseLastResponse
-			}
-			return nil
-		},
-	}, retryablehttpOptions)
-	client.CheckRetry = retryablehttp.HostSprayRetryPolicy()
-	return client
-}
-
-// makeHTTPClient creates a HTTP client with configurable redirect field
-func (r *Runner) makeDNSClientByRequest(request interface{}) *retryabledns.Client {
-	retries := r.options.Retries
-	if dnsRequest, ok := request.(*requests.DNSRequest); ok {
-		retries = dnsRequest.Retries
-	}
-
-	dnsClient, _ := retryabledns.New(DefaultResolvers, retries)
-	return dnsClient
 }
